+++ external/vulkancts/modules/vulkan/draw/vktDrawSimpleTest.cpp
@@ -152,7 +152,7 @@ tcu::TestStatus SimpleDraw::iterate (void)
 	submitCommandsAndWait(m_vk, device, queue, m_cmdBuffer.get());
 
 	// Validation
-	tcu::Texture2D referenceFrame(vk::mapVkFormat(m_colorAttachmentFormat), (int)(0.5 + WIDTH), (int)(0.5 + HEIGHT));
+	tcu::Texture2D referenceFrame(vk::mapVkFormat(m_colorAttachmentFormat), (int)(0.5f + static_cast<float>(WIDTH)), (int)(0.5f + static_cast<float>(HEIGHT)));
 
 	referenceFrame.allocLevel(0);
 
@@ -246,7 +246,7 @@ tcu::TestStatus SimpleDrawInstanced::iterate (void)
 	// Validation
 	VK_CHECK(m_vk.queueWaitIdle(queue));
 
-	tcu::Texture2D referenceFrame(vk::mapVkFormat(m_colorAttachmentFormat), (int)(0.5 + WIDTH), (int)(0.5 + HEIGHT));
+	tcu::Texture2D referenceFrame(vk::mapVkFormat(m_colorAttachmentFormat), (int)(0.5f + static_cast<float>(WIDTH)), (int)(0.5f + static_cast<float>(HEIGHT)));
 
 	referenceFrame.allocLevel(0);
